Move common apis geo + laskutus from components to utils/api. Components folder is for common general purpose svelte ui components.
