+++ src/pages/valvonta-kaytto/kohde.svelte
@@ -10,7 +10,7 @@
 
   import * as KayttajaApi from '@Pages/kayttaja/kayttaja-api';
   import * as ValvontaApi from './valvonta-api';
-  import * as geoApi from '@Component/Geo/geo-api';
+  import * as geoApi from '@Utility/api/geo-api';
 
   import Overlay from '@Component/Overlay/Overlay.svelte';
   import Spinner from '@Component/Spinner/Spinner.svelte';
